+++ src/calculateConfig.ts
@@ -10,6 +10,7 @@ import {
   DEFAULT_SCANNED_SINCE_DAYS,
   DEFAULT_VMDR_FINDING_SEVERITIES,
   DEFAULT_VMDR_FINDING_TYPES,
+  DEFAULT_VMDR_FINDING_RESULT_QIDS,
   VALID_VMDR_FINDING_TYPES,
 } from './constants';
 import { CalculatedIntegrationConfig, UserIntegrationConfig } from './types';
@@ -72,7 +73,7 @@ export function calculateConfig({
   const vmdrFindingResultQidNumbers = readPropertyAsNumberArrayFromEnvOrConfig({
     config,
     propertyName: 'vmdrFindingResultQids',
-    defaultValue: [],
+    defaultValue: DEFAULT_VMDR_FINDING_RESULT_QIDS,
   });
 
   return {