+++ src/provider/client/index.ts
@@ -186,6 +186,14 @@ export type IterateHostDetectionsOptions = {
    */
   filters?: vmpc.ListHostDetectionsFilters;
   pagination?: { limit: number };
+
+  /**
+   * Include detection results data when listing host detections. This is
+   * expensive due to the increased number of bytes transferred and processed in
+   * the XML payload.
+   */
+  includeResults?: boolean;
+
   // TODO make this a required argument and update tests
   onRequestError?: (pageIds: number[], err: Error) => void;
 };
@@ -726,63 +734,6 @@ export class QualysAPIClient {
     return assets[0];
   }
 
-  public async iterateHostDetectionsWithResults(
-    hostIds: QWebHostId[],
-    qids: number[],
-    iteratee: ResourceIteratee<{
-      host: vmpc.DetectionHost;
-      detections: vmpc.HostDetection[];
-    }>,
-    options?: IterateHostDetectionsOptions,
-  ): Promise<void> {
-    const endpoint = '/api/2.0/fo/asset/host/vm/detection/';
-
-    const filters: Record<string, string> = {};
-    if (options?.filters) {
-      for (const [k, v] of Object.entries(options.filters)) {
-        if (v && (!Array.isArray(v) || v.length > 0)) {
-          filters[k] = String(v);
-        }
-      }
-    }
-
-    const fetchDetections = async (
-      ids: QWebHostId[],
-      qids: number[],
-    ): Promise<vmpc.DetectionHost[]> => {
-      const params = new URLSearchParams({
-        ...filters,
-        action: 'list',
-        show_tags: '1',
-        show_igs: '1',
-        show_results: '1',
-        output_format: 'XML',
-        truncation_limit: String(qids.length),
-        ids: ids.map(String),
-        qids: qids.map(String),
-      });
-
-      const response = await this.executeAuthenticatedAPIRequest(
-        this.qualysUrl(endpoint),
-        { method: 'POST', body: params },
-      );
-
-      const jsonFromXml = await parseXMLResponse<
-        vmpc.ListHostDetectionsResponse
-      >(response);
-      return toArray(
-        jsonFromXml.HOST_LIST_VM_DETECTION_OUTPUT?.RESPONSE?.HOST_LIST?.HOST,
-      );
-    };
-
-    for (const host of await fetchDetections(hostIds, qids)) {
-      await iteratee({
-        host,
-        detections: toArray(host.DETECTION_LIST?.DETECTION),
-      });
-    }
-  }
-
   /**
    * Iterate detected host vulnerabilities.
    *
@@ -801,6 +752,7 @@ export class QualysAPIClient {
    *
    * @param hostIds the set of QWEB host IDs to fetch detections
    * @param iteratee receives each host and its detections
+   * @param options configure detection fetch parameters
    */
   public async iterateHostDetections(
     hostIds: QWebHostId[],
@@ -830,7 +782,7 @@ export class QualysAPIClient {
         action: 'list',
         show_tags: '1',
         show_igs: '1',
-        show_results: '0',
+        show_results: !!options?.includeResults ? '1' : '0',
         output_format: 'XML',
         truncation_limit: String(ids.length),
         ids: ids.map(String),